ENH: camflow: stagFlow/counterflow solver has been updated.
Now stagflow/counterflow solver will run, but diverges. Need
to check the code in more detail.

+++ src/camflow/include/array.h
@@ -143,7 +143,7 @@ namespace Camflow{
         protected:
 
             std::vector<doublereal> aData;
-            int nRows, nCols;
+            size_t nRows, nCols;
 
     };
